migrations/versions/dcfdd5c56c63_employee_created.py
"""Employee created

Revision ID: dcfdd5c56c63
Revises: 
Create Date: 2024-08-31 13:18:05.141176

"""
from typing import Sequence, Union

from alembic import op
import sqlalchemy as sa
import sqlmodel


# revision identifiers, used by Alembic.
revision: str = 'dcfdd5c56c63'
down_revision: Union[str, None] = None
branch_labels: Union[str, Sequence[str], None] = None
depends_on: Union[str, Sequence[str], None] = None


def upgrade() -> None:
    pass


def downgrade() -> None:
    pass
